% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeArea.R
\name{centroid}
\alias{centroid}
\title{Centroid function}
\usage{
centroid(x, ...)
}
\arguments{
\item{x}{sf polygon}

\item{...}{Not used}
}
\value{
Coordinates of the centroid of a polygon
}
\description{
Centroid function
}
