% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clipLatitude.R
\name{clipLatitude}
\alias{clipLatitude}
\title{Clip a latitude into the range -90 to 90}
\usage{
clipLatitude(latitude)
}
\arguments{
\item{latitude}{A latitude in signed decimal degrees.}
}
\value{
A latitude in signed decimal degrees between -90 and 90
}
\description{
Clip a latitude into the range -90 to 90
}
\examples{
clipLatitude(-220.3)
clipLatitude(-90.3)
clipLatitude(-90)
clipLatitude(-89.9)
clipLatitude(-22.370475875)
clipLatitude(0)
clipLatitude(24.3)
clipLatitude(89.9)
clipLatitude(90)
clipLatitude(221.3)
}
