% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeLatitudePrecision.R
\name{computeLatitudePrecision}
\alias{computeLatitudePrecision}
\title{Compute the latitude precision value for a given code length}
\usage{
computeLatitudePrecision(codeLength, grid_rows_ = 5)
}
\arguments{
\item{codeLength}{Numeric representing the length of the given open location code}

\item{grid_rows_}{Number of rows in the grid refinement method}
}
\value{
Numeric latitude precision
}
\description{
Compute the latitude precision value for a given code length. Lengths <= 10 have the same precision for latitude and longitude, but lengths > 10 have different precisions due to the grid method having fewer columns than rows.
}
\examples{
computeLatitudePrecision(8)
computeLatitudePrecision(10)
computeLatitudePrecision(11)
computeLatitudePrecision(13)
computeLatitudePrecision(14)
}
