% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode}
\alias{encode}
\title{Encode a location into an Open Location Code}
\usage{
encode(
  latitude,
  longitude,
  codeLength = 10,
  code_alphabet_ = "23456789CFGHJMPQRVWX",
  encoding_base_ = nchar(code_alphabet_),
  pair_code_length_ = 10,
  final_lat_precision_ = 2.5e+07,
  final_lng_precision_ = 8192000,
  max_digit_count_ = 15,
  latitude_max_ = 90,
  longitude_max_ = 180,
  grid_columns_ = 4,
  grid_rows_ = 5,
  grid_code_length_ = max_digit_count_ - pair_code_length_,
  seperator_position_ = 9
)
}
\arguments{
\item{latitude}{A latitude in signed decimal degrees. Will be clipped to the range -90 to 90.}

\item{longitude}{A longitude in signed decimal degrees. Will be normalised to the range -180 to 180.}

\item{codeLength}{The number of significant digits in the output code, not including any separator characters.}

\item{code_alphabet_}{The character set used to encode the values}

\item{encoding_base_}{The base to use to convert numbers to/from}

\item{pair_code_length_}{Maximum code length using lat/lng pair encoding. Excludes prefix and separator characters}

\item{final_lat_precision_}{Multiply latitude by this much to make it a multiple of the finest precision}

\item{final_lng_precision_}{Multiply longitude by this much to make it a multiple of the finest precision}

\item{max_digit_count_}{The max number of digits to process in a plus code}

\item{latitude_max_}{The maximum value for latitude in degrees}

\item{longitude_max_}{The maximum value for longitude in degrees}

\item{grid_columns_}{Number of columns in the grid refinement method}

\item{grid_rows_}{Number of rows in the grid refinement method}

\item{grid_code_length_}{Number of digits in the grid precision part of the code}

\item{seperator_position_}{Expected location of the separator character}
}
\value{
An Open Location Code
}
\description{
Produces a code of the specified length, or the default length if no length is provided. The length determines the accuracy of the code. The default length is 10 characters, returning a code of approximately 13.5x13.5 meters. Longer codes represent smaller areas, but lengths > 14 are sub-centimetre and so 11 or 12 are probably the limit of useful codes.
}
\examples{

encode(47.365590, 8.524997)
}
