% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isFull.R
\name{isFull}
\alias{isFull}
\title{Determine if a code is a valid full Open Location Code}
\usage{
isFull(
  code,
  code_alphabet_ = "23456789CFGHJMPQRVWX",
  encoding_base_ = nchar(code_alphabet_),
  latitude_max_ = 90,
  longitude_max_ = 180
)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{code_alphabet_}{The character set used to encode the values}

\item{encoding_base_}{The base to use to convert numbers to/from}

\item{latitude_max_}{The maximum value for latitude in degrees}

\item{longitude_max_}{The maximum value for longitude in degrees}
}
\value{
Logical. TRUE if a code is a valid full Open Location Code. FALSE if not.
}
\description{
Determine if a code is a valid full Open Location Code
}
\details{
Not all possible combinations of Open Location Code characters decode to valid latitude and longitude values. This checks that a code is valid and also that the latitude and longitude values are legal. If the prefix character is present, it must be the first character. If the separator character is present, it must be after four characters.
}
\examples{
isFull("8FVC9G8F+6X")
isFull("8F+6X")
isFull("9G8F+6X")
isFull("G8F+6X")
}
