% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isShort.R
\name{isShort}
\alias{isShort}
\title{Determine if a code is a valid short Open Location Code}
\usage{
isShort(code, seperator_ = "\\\\+", seperator_position_ = 9)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{seperator_}{Regular expression denoting the separator character. Usually '\\+'}

\item{seperator_position_}{Expected location of the separator character}
}
\value{
Logical. TRUE if code is a valid short Open Location Code. FALSE if not
}
\description{
Determine if a code is a valid short Open Location Code
}
\details{
A short Open Location Code is a sequence created by removing four or more digits from an Open Location Code. It must include a separator character.
}
\examples{
isShort("8FVC9GF6+6X7") # FALSE
isShort("9GF6+6X7") # TRUE
isShort("GF6+6X7") # FALSE
}
