% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValid.R
\name{isValid}
\alias{isValid}
\title{Determine if a code is valid}
\usage{
isValid(
  code,
  seperator_ = "\\\\+",
  padding_character_ = "0",
  seperator_position_ = 9,
  code_alphabet_ = "23456789CFGHJMPQRVWX"
)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{seperator_}{Regular expression denoting the separator character. Usually '\\+'}

\item{padding_character_}{Regular expression denoting the padding character. Usually '0'}

\item{seperator_position_}{Expected location of the separator character}

\item{code_alphabet_}{The character set used to encode the values}
}
\value{
Logical. TRUE if the code is a valid Open Location Code. FALSE if not
}
\description{
Determine if a code is valid
}
\details{
To be valid, all characters must be from the Open Location Code character set with at most one separator. The separator can be in any even-numbered position up to the eighth digit.
}
\examples{
isValid("8FVC9GF6+6X7")
isValid("I am not an OLC")
}
