% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeArea.R
\name{latlng.default}
\alias{latlng.default}
\title{Find the centroid of a codeArea}
\usage{
\method{latlng}{default}(x, ...)
}
\arguments{
\item{x}{codeArea object}

\item{...}{Not used}
}
\value{
codeArea centroid in latitude longitude
}
\description{
Find the centroid of a codeArea
}
