% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeArea.R
\name{new_code_area}
\alias{new_code_area}
\title{codeArea object constructor}
\usage{
new_code_area(code = character(), codeLength, polygon_sf)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{codeLength}{The number of significant digits in the output code, not including any separator characters.}

\item{polygon_sf}{sf polygon representing the bounding box of the Open Location Code}
}
\value{
codeArea object
}
\description{
codeArea object constructor
}
