% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizeLongitude.R
\name{normalizeLongitude}
\alias{normalizeLongitude}
\title{Normalize a longitude into the range -180 to 180, not including 180.}
\usage{
normalizeLongitude(longitude)
}
\arguments{
\item{longitude}{Numeric longitude}
}
\value{
Numeric longitude that is between -180 and 180
}
\description{
Normalize a longitude into the range -180 to 180, not including 180.
}
\examples{
normalizeLongitude(-200)
normalizeLongitude(-181)
normalizeLongitude(-180)
normalizeLongitude(-179)
normalizeLongitude(179)
normalizeLongitude(180)
normalizeLongitude(181)
}
