% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olc_to_sf.R
\name{olc_to_sf}
\alias{olc_to_sf}
\title{Construct a codeArea object}
\usage{
olc_to_sf(
  code,
  lat,
  lng,
  latPrecision,
  lngPrecision,
  crs_str = sf::st_crs(4326)
)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{lat}{A latitude in signed decimal degrees. Will be normalised to the range -180 to 180.}

\item{lng}{A longitude in signed decimal degrees. Will be normalised to the range -180 to 180.}

\item{latPrecision}{Multiply latitude by this much to make it a multiple of the finest precision}

\item{lngPrecision}{Multiply longitude by this much to make it a multiple of the finest precision}

\item{crs_str}{sf crs object denoting the output coordinate reference system for the sf polygon. Usually 4326}
}
\value{
codeArea object
}
\description{
Construct a codeArea object
}
