% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeArea.R
\name{print.code_area}
\alias{print.code_area}
\title{codeArea print method}
\usage{
\method{print}{code_area}(x, ...)
}
\arguments{
\item{x}{codeArea object}

\item{...}{Not used}
}
\value{
Nothing
}
\description{
codeArea print method
}
