% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recoverNearest.R
\name{recoverNearest}
\alias{recoverNearest}
\title{Recover the nearest matching code to a specified location}
\usage{
recoverNearest(
  code,
  referenceLatitude,
  referenceLongitude,
  seperator_ = "\\\\+",
  seperator_position_ = 9,
  latitude_max_ = 90
)
}
\arguments{
\item{code}{Open Location Code. Character}

\item{referenceLatitude}{The latitude (in signed decimal degrees) to use to find the nearest matching full code}

\item{referenceLongitude}{The longitude (in signed decimal degrees) to use to find the nearest matching full code}

\item{seperator_}{Regular expression denoting the separator character. Usually '\\+'}

\item{seperator_position_}{Expected location of the separator character}

\item{latitude_max_}{The maximum value for latitude in degrees}
}
\value{
The nearest full Open Location Code to the reference location that matches the short code. If the passed code was not a valid short code, but was a valid full code, it is returned with proper capitalization but otherwise unchanged.
}
\description{
Recover the nearest matching code to a specified location
}
\examples{
recoverNearest("9G8F+6X", 47.4, 8.6)
recoverNearest("8F+6X", 47.4, 8.6)
}
